feat(permission): finish can_trigger method on smart action
